refactor(arch): moved services to root src
