+++ internal/activity/event.go
